# openApi/MatcherAPI.yaml
openapi: 3.0.0
info:
  title: MatchingTargetDTO API
  version: 1.0.0
paths:
  /matcher:
    get:
      summary: Get all matched transactions
      parameters:
        - name: from
          in: query
          description: Start date (YYYY-MM-DD)
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: End date (YYYY-MM-DD)
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchedTransactionsDTO'

  /matcher/{MatchingTargetDTOId}:
    get:
      summary: Get all matched transactions
      parameters:
        - name: from
          in: query
          description: Start date (YYYY-MM-DD)
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: End date (YYYY-MM-DD)
          required: true
          schema:
            type: string
            format: date
        - name: MatchingTargetDTOId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchedTransactionsDTO'

  /matching-target:
    get:
      summary: Get all MatchingTargetDTO
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchingTargetDTO'
    post:
      summary: Create a MatchingTargetDTO
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchingTargetDTO'
      responses:
        '201':
          description: Successful creation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchingTargetDTO'
  /matching-target/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get MatchingTargetDTO by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchingTargetDTO'
        '404':
          description: MatchingTargetDTO not found
    put:
      summary: Update MatchingTargetDTO by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchingTargetDTO'
      responses:
        '200':
          description: Successful update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchingTargetDTO'
        '404':
          description: MatchingTargetDTO not found
    delete:
      summary: Delete MatchingTargetDTO by ID
      responses:
        '204':
          description: Successful deletion
        '404':
          description: MatchingTargetDTO not found
components:
  schemas:
    MatchingTargetDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: string
        iban:
          type: string
        variableSymbol:
          type: string
        specificSymbol:
          type: string
        constantSymbol:
          type: string
        reference:
          type: string
        identifier:
          type: string

    TransactionDTO:
      type: object
      required:
        - transactionId
        - date
        - amount
      properties:
        transactionId:
          type: string
          description: The transaction ID
        date:
          type: string
          format: date-time
          description: The date of the transaction
        constantSymbol:
          type: string
          description: The constant symbol associated with the transaction
        variableSymbol:
          type: string
          description: The variable symbol associated with the transaction
        accountNumberFull:
          type: string
          description: The full account number associated with the transaction
        amount:
          type: integer
          format: int64
          description: The amount of the transaction
        comment:
          type: string
          description: Any comments associated with the transaction
        currency:
          type: string
          description: The currency of the transaction
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the transaction was created

    MatchedTransactionsDTO:
      type: object
      required:
        - targetId
        - transactions
      properties:
        targetId:
          type: integer
          format: int64
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDTO'